Day 4/animal_classification/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
joblib
scikit-learn
numpy
opencv-python
jinja2
requests